backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# Anthropic SDK
anthropic==0.32.0

# LangChain + RAG
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-pinecone==0.2.0
langchain-community==0.3.1
pinecone-client==5.0.0

# Supabase
supabase==2.7.0

# Google APIs
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.148.0

# Document generation
python-pptx==1.0.2
reportlab==4.2.4
jinja2==3.1.4
pillow==11.0.0

# Utilities
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
